Revert moving transform profile to modify.
