+++ Architecture_Engine/Architecture_Engine.csproj
@@ -93,7 +93,6 @@
     <Compile Include="Modify\SetGeometry.cs" />
     <Compile Include="Modify\SetOutlineElements1D.cs" />
     <Compile Include="Modify\Transform.cs" />
-    <Compile Include="Modify\TransformProfile.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Query\Geometry.cs" />
     <Compile Include="Query\HasMergeablePropertiesWith.cs" />
